minor buf fix - using an integer value for size of kernel removed unused python file ridge_freq_2

+++ src/ridge_filter.py
@@ -77,7 +77,7 @@ def ridge_filter(im, orient, freq, kx, ky):
     sigmax = 1/unfreq[0]*kx;
     sigmay = 1/unfreq[0]*ky;
     
-    sze = np.round(3*np.max([sigmax,sigmay]));
+    sze = np.int(np.round(3*np.max([sigmax,sigmay])));
     
     x,y = np.meshgrid(np.linspace(-sze,sze,(2*sze + 1)),np.linspace(-sze,sze,(2*sze + 1)));
